requirements.txt
click==8.0.1
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
peewee==3.14.4
python-dotenv==0.17.1
six==1.16.0
Werkzeug==2.0.1
